refactor(access-service): extract LCM read-surface into access-lcm-surface.ts
God-file decomposition (#1526 playbook), third access-service cluster.
Moves lcmSearch and the read-authorization resolution trio
(resolveLcmReadNamespace, resolveRawExcerptReadNamespace,
resolveLcmReadSessionIds — the #1495/#1505 read gates) verbatim into
AccessLcmSurface (453 LOC, 9 deps via selfDeps live wiring).

Stacked on refactor/access-service-admin-ops.

access-service.ts: 7,191 -> 6,796 LOC.
